Move classes in module proxy.common to common
