+++ common/src/main/java/org/apache/rocketmq/common/utils/Start.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.proxy.common;
+package org.apache.rocketmq.common.utils;
 
 public interface Start {
     void start() throws Exception;
