+++ common/src/main/java/org/apache/rocketmq/common/utils/StartAndShutdown.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.proxy.common;
+package org.apache.rocketmq.common.utils;
 
 public interface StartAndShutdown extends Start, Shutdown {
     default void preShutdown() throws Exception {}
